add command to the prepare.py and deployer.py
